Bigram optimize

+++ src/term_bitset.rs
@@ -10,7 +10,7 @@ use std::str::FromStr;
 ///
 /// In this representation, n-th bit represents the n-th term, and
 /// is equal to 1 if it's present, and 0 if it's missing.
-pub type TermMask = u16;
+pub type TermMask = u8;
 
 fn term_mask_from_query(query: &Query, term_subset: &[Term]) -> Result<TermMask, Error> {
     let mut mask: TermMask = 0;
